requirements.txt
Django==5.0.2
djangorestframework==3.14.0
mysqlclient==2.2.4
python-dotenv==1.0.1
Pillow==10.2.0
django-crispy-forms==2.1
crispy-bootstrap5==2023.10
django-allauth==0.61.1
django-filter==24.1
drf-yasg==1.21.7
django-cors-headers==4.2.0
djangorestframework-simplejwt==5.3.1